requirements.txt
pandas
scikit-learn
seaborn
matplotlib
ipykernel
joblib
Flask
python-box
mlflow
ensure
